bugfix: ignore QueryResponse events when coalescing sequences
mysql_watcher/mysql_logger produces query logs with QueryStart and
QueryResponse events for each query.  The QueryResponse event contains
a copy of the query being responded to.  sqllog.py's CoalesceSequences
ignored the type and just adds the QueryResponse event as if it were
another query, so each query would be duplicated.  I'm not sure how this
was avoided in the generation of doc/examples/example-good-sequences.log.